tsify

+++ src/component/extrude_connector.rs
@@ -1,4 +1,7 @@
-#[derive(Debug, Clone)]
+use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::wasm_bindgen;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ExtrudeConnectorData {
     Bracket(BracketData),                 // 角码
     SlotBracket(SlotBracketData),         // 槽连接件
@@ -8,7 +11,7 @@ pub enum ExtrudeConnectorData {
     ElasticFastener(ElasticFastenerData), // 弹性扣件
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BracketData {
     series: BracketSeries,
     load: BracketLoad,
@@ -16,7 +19,7 @@ pub struct BracketData {
     manufacture_method: BracketManufactureMethod,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketSeries {
     S2020,
     S3030,
@@ -27,7 +30,7 @@ pub enum BracketSeries {
     S8080,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketLoad {
     Light(BracketLightLoadSide),
     Standard,
@@ -35,110 +38,110 @@ pub enum BracketLoad {
     Heavy,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketLightLoadSide {
     Single,
     Double,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketSurface {
     White,
     Black,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketManufactureMethod {
     Casting,                 // 铸造
     Extrusion(BracketAngle), // 挤压
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BracketAngle {
     _45,
     _135,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SlotBracketData {
     series: SlotBracketSeries,
     material: SlotBracketMaterial,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum SlotBracketSeries {
     S20,
     S30,
     S40,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum SlotBracketMaterial {
     ZincAlloy, // 锌合金
     Steel,     // 钢
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ConnectorPlateData {
     Outer(OuterConnectorPlateData),
     Inner(InnerConnectorPlateData),
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct OuterConnectorPlateData {
     plate_type: OuterConnectorPlateType,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum OuterConnectorPlateType {
     T,
     L,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct InnerConnectorPlateData {
     plate_type: InnerConnectorPlateType,
     hole: Hole,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum InnerConnectorPlateType {
     L,
     Line,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum NutData {
     Normal(NormalNutData),
     Extrude(ExtrudeNutData),
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct NormalNutData {
     nut_type: NormalNutType,
     hole: Hole,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum NormalNutType {
     FlangeNut,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ExtrudeNutData {
     extrude_nut_type: ExtrudeNutType,
     series: ExtrudeNutSeries,
     hole: Hole,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ExtrudeNutSeries {
     SW4mm,
     SW8mm,
     SW10mm,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ExtrudeNutType {
     Slide,
     T,
@@ -146,7 +149,7 @@ pub enum ExtrudeNutType {
     SpringBall,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum Hole {
     M4,
     M5,
@@ -157,44 +160,45 @@ pub enum Hole {
     M12,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum BoltData {
     Normal(NormalBoltData),
     Extrude(ExtrudeBoltData),
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct NormalBoltData {
     bolt_type: NormalBoltType,
     slot_width: u8,
     bolt_length: u16,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum NormalBoltType {
     HalfSphere,
     Plate,
     Cylinder,
 }
 
-#[derive(Debug, Clone)]
+#[wasm_bindgen]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ExtrudeBoltData {
     extrude_bolt_type: ExtrudeBoltType,
     hole: Hole,
-    bolt_length: u16,
+    bolt_length: u32, // 0.01mm
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ExtrudeBoltType {
     T,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ElasticFastenerData {
     series: ElasticFastenerSeries,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ElasticFastenerSeries {
     S30,
     S40,
